requirements.txt
pymysql
pytest
pytest-html
pyyaml
